+++ content/src/app/local-content-storage/local-content-storage.ts
@@ -137,7 +137,7 @@ export class LocalContentStorageComponent implements OnInit, OnDestroy {
 
     this.subscriptions.push(
         this.be.manifestStatus.subscribe(status => {
-          if (status === 'loaded') {
+          if (status.status === 'loaded') {
             this.be.sendRequestWithArgs('lcs-data', null) ;
           }
         })